cf_fields table field category_id
Changed from TINYINT(4) to INT(11)

+++ branches/groupoffice-6.1/www/modules/customfields/install/install.sql
@@ -49,7 +49,7 @@ CREATE TABLE IF NOT EXISTS `cf_enabled_categories` (
 -- DROP TABLE IF EXISTS `cf_fields`;
 CREATE TABLE IF NOT EXISTS `cf_fields` (
   `id` int(11) NOT NULL AUTO_INCREMENT,
-  `category_id` tinyint(4) NOT NULL DEFAULT '0',
+  `category_id` int(11) NOT NULL DEFAULT '0',
   `name` VARCHAR( 255 ) NOT NULL,
   `datatype` varchar(100) NOT NULL DEFAULT 'GO_Customfields_Customfieldtype_Text',
   `sort_index` int(11) NOT NULL DEFAULT '0',
